decompose model store internals
